back/requirements.txt
fastapi
uvicorn
transformers
torch
datasets
strawberry-graphql[fastapi]
pydantic
aiofiles
watchdog
python-multipart
fastapi-pagination